requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
lxml>=4.9.0
aiohttp>=3.9.0
pydantic>=2.5.0
python-dotenv>=1.0.0
fake-useragent>=1.4.0
undetected-chromedriver>=3.5.0
sqlalchemy>=2.0.0
aiogram>=3.3.0
gunicorn>=21.2.0
python-telegram-bot>=20.7
